feat: migrate source from JavaScript to TypeScript
- Rename all src/*.js -> src/*.ts with full type annotations
- Rename __tests__/*.test.js -> __tests__/*.test.ts
- Add tsconfig.json (strict, NodeNext, ES2022, noEmit: true)
- Add typescript-eslint flat config (syntactic rules, no project reference)
- Add @rollup/plugin-typescript to rollup.config.js (keep as .js)
- Add tsx devDep for test runner; update test script to use tsx/esm loader
  to remap .js -> .ts imports at runtime (node --strip-types does not do this)
- Add typecheck script (tsc --noEmit); add to all and CI
- Fix process.env.RUNNER_TEMP! non-null assertions in convert.ts,
  setup-dotnet.ts, install-tool.ts
- Rename CI job test-javascript -> test-typescript; add Typecheck step
- Rebuild dist bundles

devDependencies added: typescript, @rollup/plugin-typescript, @types/node,
tsx, typescript-eslint
