+++ .github/workflows/ci.yml
@@ -12,8 +12,8 @@ permissions:
   contents: read
 
 jobs:
-  test-javascript:
-    name: JavaScript Tests
+  test-typescript:
+    name: TypeScript Tests
     runs-on: ubuntu-latest
 
     steps:
@@ -35,6 +35,10 @@ jobs:
       - name: Lint
         id: npm-lint
         run: npm run lint
+
+      - name: Typecheck
+        id: npm-typecheck
+        run: npm run typecheck
 
       - name: Test
         id: npm-test
