+++ eslint.config.js
@@ -1,23 +1,12 @@
-import js from '@eslint/js';
+import tseslint from 'typescript-eslint';
 
-export default [
+export default tseslint.config(
+  { ignores: ['dist/**', 'node_modules/**'] },
+  ...tseslint.configs.recommended,
   {
-    ignores: ['dist/**', 'node_modules/**']
-  },
-  {
-    files: ['src/**/*.js', '__tests__/**/*.js'],
-    languageOptions: {
-      ecmaVersion: 2024,
-      sourceType: 'module',
-      globals: {
-        console: 'readonly',
-        process: 'readonly',
-        Buffer: 'readonly'
-      }
-    },
+    files: ['src/**/*.ts', '__tests__/**/*.ts'],
     rules: {
-      ...js.configs.recommended.rules,
-      'no-unused-vars': ['error', { argsIgnorePattern: '^_' }]
+      '@typescript-eslint/no-unused-vars': ['error', { argsIgnorePattern: '^_' }]
     }
   }
-];
+);
