+++ package.json
@@ -11,14 +11,15 @@
     "package:watch": "npx rollup --config rollup.config.js --watch",
     "build": "npm run package",
     "lint": "eslint src/ __tests__/ --max-warnings 0",
-    "test": "node --test",
-    "all": "npm run lint && npm run test && npm run package"
+    "typecheck": "tsc --noEmit",
+    "test": "node --import tsx/esm --test __tests__/**/*.test.ts",
+    "all": "npm run lint && npm run typecheck && npm run test && npm run package"
   },
   "dependencies": {
+    "@actions/artifact": "^2.3.0",
     "@actions/core": "^3.0.0",
     "@actions/exec": "^1.1.1",
     "@actions/glob": "^0.5.0",
-    "@actions/artifact": "^2.3.0",
     "@actions/io": "^1.1.3",
     "@actions/tool-cache": "^2.0.1"
   },
@@ -26,9 +27,14 @@
     "@rollup/plugin-commonjs": "^29.0.0",
     "@rollup/plugin-json": "^6.1.0",
     "@rollup/plugin-node-resolve": "^16.0.3",
+    "@rollup/plugin-typescript": "^12.3.0",
+    "@types/node": "^25.9.1",
     "eslint": "^9.0.0",
+    "rimraf": "^6.0.0",
     "rollup": "^4.57.1",
-    "rimraf": "^6.0.0"
+    "tsx": "^4.22.3",
+    "typescript": "^6.0.3",
+    "typescript-eslint": "^8.60.0"
   },
   "engines": {
     "node": ">=24"
