+++ rollup.config.js
@@ -1,21 +1,27 @@
 import commonjs from '@rollup/plugin-commonjs';
 import { nodeResolve } from '@rollup/plugin-node-resolve';
 import json from '@rollup/plugin-json';
+import typescript from '@rollup/plugin-typescript';
 
 const shared = {
   output: { esModule: true, format: 'es', sourcemap: true },
-  plugins: [commonjs(), nodeResolve({ preferBuiltins: true }), json()],
+  plugins: [
+    typescript(),
+    commonjs(),
+    nodeResolve({ preferBuiltins: true }),
+    json(),
+  ],
 };
 
 export default [
   {
     ...shared,
-    input: 'src/index.js',
+    input: 'src/index.ts',
     output: { ...shared.output, file: 'dist/index.js' },
   },
   {
     ...shared,
-    input: 'src/post.js',
+    input: 'src/post.ts',
     output: { ...shared.output, file: 'dist/post.js' },
   },
 ];
