+++ src/index.ts
@@ -1,0 +1,3 @@
+import { run } from './main.js';
+
+run();
