--- src/post.js
@@ -1,3 +1,0 @@
-import { cleanup } from './cleanup.js';
-
-cleanup();
