+++ src/post.ts
@@ -1,0 +1,3 @@
+import { cleanup } from './cleanup.js';
+
+cleanup();
